refactor: organize code for migration cli
